Expose health check failures
We use the healthz package from k8s.io/apiserver to implement our health
check. However, that check does not expose the failure reason in the
logs. Example output:

I1002 08:05:39.882128       1 healthz.go:200] [-]godo failed: reason withheld

As described in an upstream code comment [1], the rationale here is that the
endpoint is public and not supposed to give away sensitive data.
Instead, Kubernetes users are expected to inspect failure details
through separate, authenticated access [2].

This makes the package a bad fit for our use case because (1) we do not
have an alternative way to inspect errors and (2) we do not need to:
users concerned about the information that is being returned from the
health check implementation (currently the respone from the /v2/account
endpoint of the DigitalOcean API) can configure CCM's debug endpoint to
run on localhost.

Consequently, we re-implement the health check as a custom, plain
net/http handler and log health check failures explicitly.

[1]: https://github.com/digitalocean/digitalocean-cloud-controller-manager/blob/f3ea5608898f253f7bfd85de35f166dde36a278a/vendor/k8s.io/apiserver/pkg/server/healthz/healthz.go#L226-L227
[2]: https://stackoverflow.com/a/53861547